rename util to load-external-script
